init @vencord/discord-types package
