+++ package.json
@@ -53,6 +53,7 @@
         "@types/react": "^19.0.10",
         "@types/react-dom": "^19.0.4",
         "@types/yazl": "^2.4.5",
+        "@vencord/discord-types": "link:packages/discord-types",
         "diff": "^7.0.0",
         "discord-types": "^1.3.26",
         "esbuild": "^0.25.1",
