+++ packages/discord-types/classes.d.ts
@@ -1,0 +1,21 @@
+export interface ImageModalClasses {
+    image: string,
+    modal: string,
+}
+
+export interface ButtonWrapperClasses {
+    hoverScale: string;
+    buttonWrapper: string;
+    button: string;
+    iconMask: string;
+    buttonContent: string;
+    icon: string;
+    pulseIcon: string;
+    pulseButton: string;
+    notificationDot: string;
+    sparkleContainer: string;
+    sparkleStar: string;
+    sparklePlus: string;
+    sparkle: string;
+    active: string;
+}
