+++ packages/discord-types/index.d.ts
@@ -1,0 +1,6 @@
+export * as Classes from "./classes";
+export * as Components from "./components";
+export * as FluxEvents from "./fluxEvents";
+export * as Menu from "./menu";
+export * as Stores from "./stores";
+export * as Utils from "./utils";
