+++ packages/discord-types/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "@vencord/discord-types",
+  "author": "Vencord Contributors",
+  "description": "Typescript definitions for the webpack modules of the Discord Web app",
+  "version": "1.0.0",
+  "license": "LGPL-3.0-or-later",
+  "types": "src/index.d.ts",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/Vendicated/Vencord.git"
+  },
+  "dependencies": {
+    "@types/react": "^19.0.10",
+    "type-fest": "^4.41.0"
+  }
+}
