+++ src/webpack/common/classes.ts
@@ -16,9 +16,8 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 
+import * as t from "@vencord/discord-types/classes";
 import { findByPropsLazy, findLazy } from "@webpack";
-
-import * as t from "./types/classes";
 
 export const ModalImageClasses: t.ImageModalClasses = findLazy(m => m.image && m.modal && !m.applicationIcon);
 export const ButtonWrapperClasses: t.ButtonWrapperClasses = findByPropsLazy("buttonWrapper", "buttonContent");
