+++ src/webpack/common/index.ts
@@ -21,8 +21,5 @@ export * from "./components";
 export * from "./menu";
 export * from "./react";
 export * from "./stores";
-export * as ComponentTypes from "./types/components.d";
-export * as MenuTypes from "./types/menu.d";
-export * as UtilTypes from "./types/utils.d";
 export * from "./userSettings";
 export * from "./utils";
