+++ src/webpack/common/menu.ts
@@ -16,9 +16,8 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 
+import type * as t from "@vencord/discord-types/menu";
 import { filters, mapMangledModuleLazy, waitFor, wreq } from "@webpack";
-
-import type * as t from "./types/menu";
 
 export const Menu = {} as t.Menu;
