+++ src/webpack/types.d.ts
@@ -1,0 +1,14 @@
+/*
+ * Vencord, a Discord client mod
+ * Copyright (c) 2025 Vendicated and contributors
+ * SPDX-License-Identifier: GPL-3.0-or-later
+ */
+
+// TODO: figure out a way to not have this and still be backwards compatible for userplugins
+
+export * from "@vencord/discord-types/classes";
+export * from "@vencord/discord-types/components";
+export * from "@vencord/discord-types/fluxEvents";
+export * from "@vencord/discord-types/menu";
+export * from "@vencord/discord-types/stores";
+export * from "@vencord/discord-types/utils";
