+++ tsconfig.json
@@ -27,7 +27,7 @@
             "@components/*": ["./components/*"],
             "@utils/*": ["./utils/*"],
             "@shared/*": ["./shared/*"],
-            "@webpack/types": ["./webpack/common/types"],
+            "@webpack/types": ["./webpack/types.d.ts"],
             "@webpack/common": ["./webpack/common"],
             "@webpack": ["./webpack/webpack"],
             "@webpack/patcher": ["./webpack/patchWebpack"],
